refactor: move example files
