+++ example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/search/EmbeddedDocumentSearchRepository.kt
@@ -1,4 +1,4 @@
-package com.github.inflab.example.spring.data.mongodb.search.embeddedDocument
+package com.github.inflab.example.spring.data.mongodb.search
 
 import com.github.inflab.example.spring.data.mongodb.entity.supplies.SaleItem
 import com.github.inflab.example.spring.data.mongodb.entity.supplies.Sales
