+++ example/spring-data-mongodb/src/main/kotlin/com/github/inflab/example/spring/data/mongodb/search/PhraseSearchRepository.kt
@@ -1,4 +1,4 @@
-package com.github.inflab.example.spring.data.mongodb.search.phrase
+package com.github.inflab.example.spring.data.mongodb.search
 
 import com.github.inflab.example.spring.data.mongodb.entity.mflix.Movies
 import com.github.inflab.spring.data.mongodb.core.aggregation.aggregation
